Support custom excerpt for posts
We might want to control what goes in the excerpt displayed for a post
on the index page, for example to sanitize HTML layouts and customize
length etc.
This patch checks for variable `post.custom_excerpt` for each post in
the post flow and uses that instead of the content if provided.